feat(conduct): refactor with Contributor Covenant v3
This is a full rewrite based on the latest version of the Contributor Covenant. The important thing is that the spirit of the rules is still the same. Most of the changes here are just cleaner wording and organization. It is worth noting that the first two steps of the enforcement ladder have changed, switching from "correction" and "warning" to "warning" and "temporarily limited activities."